minor: switch total time-limit to doubles
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@14246 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ blind/blind-main.c
@@ -163,7 +163,7 @@ static int read_parameters(blind_t* bp) {
 		} else if (is_word(line, "timelimit ", &nextword)) {
 			bp->timelimit = atoi(nextword);
 		} else if (is_word(line, "total_timelimit ", &nextword)) {
-			bp->total_timelimit = atoi(nextword);
+			bp->total_timelimit = atof(nextword);
 		} else if (is_word(line, "total_cpulimit ", &nextword)) {
 			bp->total_cpulimit = atof(nextword);
 		} else if (is_word(line, "verify_pix ", &nextword)) {
